requirements-api.txt
# Dependencias minimas para la API en Docker
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
imbalanced-learn>=0.11
fastapi>=0.115
uvicorn>=0.30
pydantic>=2.0
joblib>=1.4
evidently>=0.7
shap>=0.45
google-genai>=1.0
anthropic>=0.40
python-dotenv>=1.0
